+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.0.0-rc.1] - 2026-06-27
 
 ### Added
-
+- Search filters are now preserved across collection and item searches
 - Adding `extent`s to the root catalog will restrict the Search filters
 - Support free-text search for Collections in list of collections
 - Add a link to Collection Search from the Collections overview page for advanced filters